Move TemplateModel into metamodel
