+++ tests/test_gromet.py
@@ -5,7 +5,8 @@ from pathlib import Path
 from dataclasses import asdict
 from mira.modeling.gromet_model import GrometModel, model_to_gromet_json_file, model_to_gromet
 from mira.metamodel import ControlledConversion, Concept, NaturalConversion
-from mira.modeling import Model, TemplateModel
+from mira.modeling import Model
+from mira.metamodel.templates import TemplateModel
 
 
 def _get_sir_model_templ():
