+++ tests/test_model_api.py
@@ -11,7 +11,8 @@ from fastapi.testclient import TestClient
 
 from mira.dkg.model import model_blueprint#, ToGrometQuery
 from mira.metamodel import Concept, ControlledConversion, NaturalConversion
-from mira.modeling import TemplateModel, Model
+from mira.modeling import Model
+from mira.metamodel.templates import TemplateModel
 # from mira.modeling.gromet_model import GrometModel
 from mira.modeling.ops import stratify
 from mira.modeling.petri import PetriNetModel
